Create intersectionOfTwoArrays.java
It is one of the best question related to hashset and build strong foundation in it. This Java code finds and prints the intersection of two arrays, arr1 and arr2, using HashSet. It does so by creating two HashSet objects, set1 and set2, to store unique elements from each array. Then, it iterates through set1, checking if each element is also present in set2. Common elements are added to a third HashSet called result, which holds the intersection. Finally, the code prints the elements in the result set, which represent the common elements between the two arrays.